simplify package overrides
Previously we had four different files for package overrides and
each manifest list pulled in from one of the files (namespaced to
the target OS/Version). With conditional includes we can make this
smarter.

Move to a model where we have a single packages.overrides.yaml file
that conditionally includes overrides based on the OS/Version and
unconditionally include that in common.yaml.