passwd,group: sync with rhel-bootc
This imports all the entries from the passwd/group of rhel-bootc. The
latter is mostly a superset of what we already had here. Only two groups
have conflicts: `tape` and `input`. The first changes GID from 30 to 33,
and the second from 798 to 104. Those are the more canonical GIDs as per
https://src.fedoraproject.org/rpms/setup/blob/rawhide/f/uidgid.

Those GIDs have been changed as well in FCOS:
- https://github.com/coreos/fedora-coreos-config/pull/1832.
- https://github.com/coreos/fedora-coreos-config/pull/1866.

Changing them should be quite safe. They are normally only used to own
device nodes in devtmpfs rather than actual files on disk.

This would happen as part of the rebase to rhel-bootc anyway, but doing
it separately allows me to minimize the actual diff during the rebase
while also surfacing the important changes in separate commits.