Build via Containerfile and derive from rhel-bootc
This builds RHEL CoreOS using a Containerfile. It uses the [FROM scratch
flow](https://docs.fedoraproject.org/en-US/bootc/building-from-scratch/)
to do a base compose.

This is structured in a way that we can build FCOS using _both_ `podman
build` _and_ `cosa build`. This allows us to make the cutover much
smoother.

This builds on top of
https://github.com/coreos/fedora-coreos-config/pull/3348. Notably, the
Containerfile and build-rootfs files are shared with FCOS. See that PR
message for more details on the overall approach since it applies as
well here.