manifest-el9-shared: apply dracut fix in postcript
This is a follow-up of [1], we have to apply the dracut fix here as
the new dracut release containing the same fix landed in F42 and EL10
repos. More details in [2].
Also, we apply part of the patch [3] to skip default
multipath.conf and let mpathconf from device-mapper-multipath do the
job.

[1] https://github.com/coreos/fedora-coreos-config/pull/3588
[2] https://github.com/coreos/fedora-coreos-tracker/issues/1937
[3] https://github.com/dracut-ng/dracut-ng/pull/1259