prow-entrypoint: always prepare repos in cosa_init
PR #44 moved prepare_repos to cosa_init because 1. we want to take
advantage of the $variant variable defined in this function and not
have to define another time in cosa_build 2. it feel the right
place to fetch the repos during the cosa init operation.
But, as prepare_repos is placed after testing the existence of the
'builds' directory, the repos are not fetch when already in an
intilialized cosa dir.
This changes the behavior by always fetching the repos during the
cosa_init process.