drop use of 25azure-udev-rules f-c-c overlay for el9
This was removed from the fedora-coreos-config upstream [1]
because the azure-vm-utils package now provides these rules.
The azure-vm-utils package is available in RHEL 9.8, which
we're now using as our development RHEL 9 version so we can
drop the use of the overlay here for el9 and unify with how
we are operating for el10 (relying on the azure-vm-utils RPM).

[1] https://github.com/coreos/fedora-coreos-config/commit/02f7e0efb07c6f6b2a4ae7ccdbd443b11bfa4e0e